Restructure testing directory
