+++ .gitignore
@@ -54,3 +54,4 @@ yarn-error.log*
 *todo.txt*
 *minio.sys*
 *.tar.gz
+.mypy_cache
