+++ src/test/cli/test_cmd_version.py
@@ -3,8 +3,8 @@
 
 import pkg_resources
 
-import common
-import cli.utils as utils
+from test import common
+from test.cli import utils
 
 from inspect import currentframe
 from types import FrameType
