+++ src/test/cli/test_misc.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
-import common
-import cli.utils as utils
+from test import common
+from test.cli import utils
 
 from inspect import currentframe
 from types import FrameType
