+++ src/test/lib/json/file-group-provider.json
