+++ src/test/lib/json/resource-groups.json
