+++ src/test/lib/json/session-property-manager.json
