+++ src/test/pyproject.toml
@@ -15,6 +15,5 @@ dependencies = [
   "requests>=2.32.2"
 ]
 
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["src.cli", "src.lib"]
+[tool.setuptools]
+package-dir = {"" = "."}
